rename: 스타일 파일 폴더 분리
