+++ src/components/test/TestCard/TestCard.tsx
@@ -1,0 +1,26 @@
+import tw from "tailwind-styled-components";
+import { TestCardProps } from "@/@types/index";
+import * as S from "./TestCard.styles";
+
+const TestCard = ({
+  answer,
+  index,
+  onClick,
+  animate,
+  animationStart
+}: TestCardProps) => {
+  const onClickCard = onClick(index);
+  return (
+    <S.Testcard
+      onClick={() => {
+        onClickCard();
+        animationStart();
+      }}
+      key={animate.toString()}
+    >
+      <S.TestcardText>{answer}</S.TestcardText>
+    </S.Testcard>
+  );
+};
+
+export default TestCard;
