+++ src/components/test/TestQuestion/TestQuestion.styles.tsx
@@ -1,0 +1,28 @@
+import tw from "twin.macro";
+
+export const QuestionDiv = tw.div`
+mb-9
+h-[190px]
+
+select-none
+relative
+animate-testComponent
+`;
+
+export const QuestionNum = tw.h2`
+text-center
+text-3xl
+font-black
+text-violet-400
+mb-3
+mt-10
+`;
+
+export const QuestionHeaderText = tw.h3`
+text-center 
+w-80 
+text-xl 
+font-semibold 
+text-slate-50 
+mb-11
+`;
