+++ src/components/test/TestQuestion/TestQuestion.tsx
@@ -1,0 +1,12 @@
+import tw from "tailwind-styled-components";
+import { QuestionText } from "@/@types/index";
+import * as S from "./TestQuestion.styles";
+
+export default function TestQuestion({ idx, subject, animate }: QuestionText) {
+  return (
+    <S.QuestionDiv>
+      <S.QuestionNum>{idx}</S.QuestionNum>
+      <S.QuestionHeaderText>{subject}</S.QuestionHeaderText>
+    </S.QuestionDiv>
+  );
+}
